TOOLS: move `ReplParser` to `tools-repl`
